tts/requirements.txt
google-cloud-texttospeech
boto3
mutagen
python-pptx
moviepy
pdf2image
pillow